Add changes

+++ kylieq-CIS415F19-P2/part3.c
@@ -69,7 +69,7 @@ void sigalrm_handler(int sig_num) {
 
     //printf("Process: %d Counter: %d\n", PCBS[COUNTER]->pid, COUNTER);
     while(1) {
-        if (PCBS[COUNTER]->state == RUNNING) {
+        if (PCBS[COUNTER]->state == RUNNING && PCBS[COUNTER]->exit_status != 1) {
             kill(PCBS[COUNTER]->pid, SIGSTOP);
             printf("Process: %d - Received Signal SIGALRM - Suspended\n", PCBS[COUNTER]->pid);
             PCBS[COUNTER]->state = STOPPED;
@@ -82,7 +82,7 @@ void sigalrm_handler(int sig_num) {
     }
 
     while(1) {
-        if (PCBS[COUNTER]->state == STOPPED) {
+        if (PCBS[COUNTER]->state == STOPPED && PCBS[COUNTER]->exit_status != 1) {
             kill(PCBS[COUNTER]->pid, SIGCONT);
             printf("Process: %d - Received Signal SIGALRM - Continued\n", PCBS[COUNTER]->pid);
             PCBS[COUNTER]->state = RUNNING;
